split doc values in proper sub formats
